Revert "Develop"

+++ back/src/main/java/com/mongoosereum/dou_survey_zone/api/v1/domain/tag/SurveyTag.java
@@ -9,5 +9,5 @@ import lombok.Data;
 @Builder
 public class SurveyTag {
     private String _id;
-    private Integer Tag_ID;
+    private Long Tag_ID;
 }
